% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{find.shape}
\alias{find.shape}
\title{Measures the BB width, BB length, perimeter, convex perimeter, area, convex hull area, compactness, roundness, and maximum length values for each object in the image.}
\usage{
find.shape(image, pix.min = 4000, scaledown = 4, colorcard = "bottomright")
}
\arguments{
\item{image}{The image to be processed, must be a JPEG}

\item{pix.min}{is number of pixels expected for the smallest object to be designated a tuber}

\item{scaledown}{by which image is divided for faster computing (a reduction in image size)}

\item{colorcard}{= NULL/"bottomright" to remove a color card if used}
}
\value{
A nested list with the BB width, BB length, perimeter, convex perimeter, area, convex hull area, compactness, roundness, and maximum length values for each object.
}
\description{
Measures the BB width, BB length, perimeter, convex perimeter, area, convex hull area, compactness, roundness, and maximum length values for each object in the image.
}
\examples{
  find.shape(system.file("images", "2020fy2_317.jpg", package = "TubAR"))
  find.shape(system.file("images", "2020fy2_317.jpg", package = "TubAR"), pix.min=4e4, scaledown=5, colorcard="bottomright")
}
