% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skin_func.R
\name{find.skin}
\alias{find.skin}
\title{Find skinning threshold across the image and return skinning percent by tuber
assumes all potatoes in the image are a sampling of the same clone (:. will have same difference between skin/flesh colors)}
\usage{
find.skin(
  image,
  display = T,
  mode = "debug",
  write.clean = F,
  pix.min = 40000,
  scaledown = 8,
  colorcard = "bottomright",
  n.core = 2,
  color_correct = T
)
}
\arguments{
\item{image}{The image to be processed, must be a JPEG}

\item{display}{=T plots to quartz}

\item{mode}{="debug" plots skinned area, numbers each tuber}

\item{write.clean}{=T saves a 'clean_image.jpg' with background whited out (for presentations, etc)}

\item{pix.min}{is number of pixels expected for the smallest object to be designated a tuber}

\item{scaledown}{by which image is divided for faster computing (a reduction in image size)}

\item{colorcard}{= NULL/"bottomright" to remove a color card if used}
}
\value{
A list containing the redness, skinning, and lightness of each object in the image
}
\description{
Find skinning threshold across the image and return skinning percent by tuber
assumes all potatoes in the image are a sampling of the same clone (:. will have same difference between skin/flesh colors)
}
\examples{
  find.skin(system.file("images", "2020fy2_317.jpg", package = "TubAR"))
  find.skin(system.file("images", "2020fy2_317.jpg", package = "TubAR"),display=T, mode="debug", write.clean=F, pix.min=4e4, scaledown=7, colorcard="bottomright", n.core=2, color_correct=T)
}
