% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{get.perim}
\alias{get.perim}
\title{Calculates the perimeter length from the coordinates of an object outline}
\usage{
get.perim(xy)
}
\arguments{
\item{xy}{An integer vector of points on the convex hull of an object}
}
\value{
The perimeter length of the convex hull in pixels
}
\description{
Calculates the perimeter length from the coordinates of an object outline
}
