% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{get.trace}
\alias{get.trace}
\title{Returns the outline of an image object.}
\usage{
get.trace(Morig)
}
\arguments{
\item{Morig}{A binarized image matrix containing a single object}
}
\value{
Binarized image matrix of the outline of an object
}
\description{
Returns the outline of an image object.
}
