% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Gives the length and width of the bounding box that would contain an object.}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{A matrix of object outline coordinates}
}
\value{
A list of the bounding box endpoint coordinates, width, height, and angle
}
\description{
Gives the length and width of the bounding box that would contain an object.
}
