% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_func.R
\name{grabcard}
\alias{grabcard}
\title{get color card, segment and return the 24 colors as RGB.}
\usage{
grabcard(image, colorcard = "bottomright", scaledown = 8, pix.min = 40000)
}
\arguments{
\item{image}{Input JPEG image}

\item{colorcard}{Declares which corner the color card is in. "bottomright" by default. Can also be set to "bottomleft", "topright", and "topleft"}

\item{scaledown}{The amount image sizes will be reduced in order to aid processing speed}

\item{pix.min}{The minimum pixel size required for objects to not be removed with the background,}
}
\value{
An array of the 24 colors as RGB
}
\description{
get color card, segment and return the 24 colors as RGB.
}
