% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{shape.all}
\alias{shape.all}
\title{Perform `find.shape` for all JPEG files in the working directory.}
\usage{
shape.all(n.core = 2)
}
\arguments{
\item{n.core}{The number of processor cores to use in processing. Default is 2.}
}
\value{
A dataframe containing the image,BB width, BB length, perimeter, convex perimeter, area, convex hull area, compactness, roundness, and maximum length values for each object in all images in the working directory.
}
\description{
Perform `find.shape` for all JPEG files in the working directory.
}
\examples{
  setwd(system.file("images", package = "TubAR"))
  shape.all(n.core=1)
}
