% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{skin.export}
\alias{skin.export}
\title{Turn the output from skin.all into a matrix of median trait values}
\usage{
skin.export(color_list, remove_tag = TRUE, green_thresh = -5, red_thresh = 50)
}
\arguments{
\item{color_list}{Output from skin.all function}

\item{remove_tag}{If TRUE, objects below `green_thresh` or above `red_thresh` will be removed from the data before taking median values}

\item{green_thresh}{Minimum CIELAB green-red value for object data to not be removed}

\item{red_thresh}{Maximum CIELAB green-red value for object data to not be removed}
}
\value{
A matrix of the median values of redness, skinning, and lightness for each image
}
\description{
Turn the output from skin.all into a matrix of median trait values
}
\examples{
  load(system.file("example_data", "skin_exmp.rds", package = "TubAR"))
  skin.export(skin_exmp, remove_tag=TRUE, green_thresh=-5, red_thresh=50)
}
