% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_func.R
\name{unpack}
\alias{unpack}
\title{Unpack shape.all into a dataframe in melted format.}
\usage{
unpack(results)
}
\arguments{
\item{results}{Input from shape.all}
}
\value{
A dataframe in melted format
}
\description{
Unpack shape.all into a dataframe in melted format.
}
